Modify: divide to presentational and containing component
