+++ src/components/ContentList/ContentList.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { connect } from "react-redux";
 import { Button, Form, Row, Col } from "react-bootstrap";
 
 const ContentList = ({
@@ -72,27 +71,4 @@ const ContentList = ({
   );
 };
 
-const mapStateToProps = (state) => ({
-  chapters: state,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  toggleReady: (idx) =>
-    dispatch({
-      type: "TOGGLE_CHAPTER_READY",
-      idx,
-    }),
-  toggleSubSectionReady: (idx, sectionIdx) =>
-    dispatch({
-      type: "TOGGLE_SUBSECTION_READY",
-      idx,
-      sectionIdx,
-    }),
-  addChapter: (title) =>
-    dispatch({
-      type: "ADD_CHAPTER",
-      title,
-    }),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(ContentList);
+export default ContentList;
